refactor: migrate coordinator to package
