+++ custom_components/thessla_green_modbus/coordinator/__init__.py
@@ -1,0 +1,15 @@
+"""Coordinator package public API."""
+
+from .coordinator import (
+    CoordinatorConfig,
+    ThesslaGreenModbusCoordinator,
+    _PermanentModbusError,
+    get_register_definition,
+)
+
+__all__ = [
+    "CoordinatorConfig",
+    "ThesslaGreenModbusCoordinator",
+    "_PermanentModbusError",
+    "get_register_definition",
+]
