+++ custom_components/thessla_green_modbus/coordinator/errors.py
